CBOR: implement bytestring parsing functions
This change implements bytestring parsing functions most of these are
defined in section D.5 in the plutus spec. ByteStrings are the underlying
encoding of the constructors `I Integer` where the size of `Integer`
exceeds 64 bits and `B ByteString` of the plutus `Data` ADT. Subsequent
commits will use these functions to parse these constructors.