CBOR: fix decoder for DData
Previous commits implementing the branching for DData caused many issues
because the branching was largely based on matching on the result of
DHead which ended up parsing beyond the input bytes and resulted in
buffer overflow runtime errors. Doing this resulted in several incorrect
matches.

This change fixes this by matching DData branches on the first byte which
is called `HeadByte`. This also tightens up several computations such as
computing `m` for indefinite length data.

Also, this change fixes several errors resulting from incorrect
implementations in the `List` constructor decoder.