Implement CBOR decoder (#273)
* CBOR: implement and test decoders d and DHead

These functions are listed in section D.4 of the plutus specification
and are core functions used to parse CBOR terms. Subsequent commits
will use these functions to parse CBOR-encoded data constants and
translate them to the TextualData sort.

The UPLC-CBOR-PARSER module was imported in uplc-configuration.md to
facilitate unit testing.

* CBOR: implement decoder for positive integers

* Configuration: translate CBOR data to textualData

To improve readability of CBOR-encoded data constnats, convert them
to the textualData format after the input program has been parsed.
By doing so, the input program is translated to a human readable format
and the semantics of builtin functions that operate on data can
continue to use the textual format.

* add test to decode small integers

* CBOR: implement negative integer decoding

Negative integers are encoded with a major type 1 to indicate negative
integers. Aside from this, the data is encoded in the same manner.

* Adding a negative integer test

* trivial: fix spacing

* CBOR: implement bytestring parsing functions

This change implements bytestring parsing functions most of these are
defined in section D.5 in the plutus spec. ByteStrings are the underlying
encoding of the constructors `I Integer` where the size of `Integer`
exceeds 64 bits and `B ByteString` of the plutus `Data` ADT. Subsequent
commits will use these functions to parse these constructors.

* CBOR: decode Integers that span over 64 bits

* Extract function DZ from DData

DZ is used in several other functions. Extract it so that it may be used
in the `DCtag` function (to be implemented in the future).

* trivial: edit comments

* CBOR: decode ByteString data

The bytestring decoder is invoked from DData by matching on the major
type and the `M` in the expression `M = ( N -Int 31 ) /32` where `N` is
the first byte of the input CBOR encoding.

* Add CBOR bytestring tests

* CBOR: refactor DData to work with BitStreams

Eventually, DData will be called recursively but functions that parse
the constructors `List`, `Map`, and `Constr i` so this function will
decode starting at the middle of the input CBOR encoding. Pass
bitstream as a parameter instead of bytes.

* trivial: rename BYTESTRING_TERMINATOR

This identifier indicates a value that represents a terminator for
any indefinite-length data and is not limited to bytestrings.

* CBOR: implement decoders for the List constructor

* CBOR: add decoder for the Map constructor

* CBOR: add decoder for Constr Integer [Data] constructor

* Integrate list, constr, and map decoders to DData

Note: it's difficult to write uplc code to test these individually so
future commits will add tests.

* Clarify rules by adding owise and requires clauses

There were several owise and requires clauses left out of rules. This
results in several overlapping rules. Add the clauses to ensure that
rules apply correctly.

* CBOR: fix decoder for DData

Previous commits implementing the branching for DData caused many issues
because the branching was largely based on matching on the result of
DHead which ended up parsing beyond the input bytes and resulted in
buffer overflow runtime errors. Doing this resulted in several incorrect
matches.

This change fixes this by matching DData branches on the first byte which
is called `HeadByte`. This also tightens up several computations such as
computing `m` for indefinite length data.

Also, this change fixes several errors resulting from incorrect
implementations in the `List` constructor decoder.

* Add unit tests for list, map, constr, and a large example

* Fix requires clause in Constr decoder

Constr decoder was being called on the llvm backend for large negative
integers. Adding these clauses distinguishes Constr and large Integer
CBOR encodings.

* Add tests for lists, maps, constr, and integers

These were derived from the unit tests that exercises the CBOR decoder

* Use overloaded DHead to eliminate code duplication

* trivial: add explanation to terminator

* change owise to a requires clause

owise will lower the priority of the rule but what's intended is to rule
out a specific case for the first parameter.

* Edit comments and fix trivial stylistic errors

Co-authored-by: Christiano Braga <christiano.braga@runtimeverification.com>
Co-authored-by: rv-jenkins <admin@runtimeverification.com>