flat parser: add support for parsing data constants (#328)
* flat parser: fix comment describing builtins

Builtins only span 7 bits. Previous specification listed 8 bits but this
was a mistake.

* flat parser: edit identifiers, no functional change

This change is strictly for readability. When reading the code it's
helpful to understand that the macros represent types. Therefore, append
"_TYPE" to each macro identifier that represent types.

* flat parser: add support for parsing data constants

According to the spec, data constants are encoded in the same way as
bytestrings constants.

* flat parser: Add unit tests for data constants

Co-authored-by: rv-jenkins <admin@runtimeverification.com>